# app/src/main/resources/openapi/documentation.yaml
openapi: 3.0.3
info:
  title: Universe API
  description: |
    REST and WebSocket API for the Universe orchestrator.
    
    ## Authentication
    The API uses Bearer token authentication. Include your token in the `Authorization` header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```
    
    ## Permissions
    - **ALL** — Full access to all endpoints (admin, instance management, config changes)
    - **PUBLIC** — Read-only access to public endpoints (instance info, logs, live-log)
    
    ## Rate Limiting
    Public API keys are rate-limited (default: 100 calls per 10 seconds). 
    Admin keys (ALL permission) bypass rate limits entirely.
    
    ## WebSocket Endpoints
    WebSocket connections require Bearer token authentication during the handshake.
    For browser clients that cannot set custom headers, pass the token as a query parameter:
    ```
    ws://host/api/console?token=YOUR_API_KEY
    ws://host/api/instances/{id}/live-log?token=YOUR_API_KEY
    ```
  version: 1.0.0
  contact:
    name: Universe Team
servers:
  - url: http://localhost:7000
    description: Local development server (default apiPort)
paths:
...
  /console:
    get:
      summary: Interactive console (WebSocket)
      description: |
        Opens an interactive WebSocket console to the Universe master node.
        
        Connect with:
        ```
        ws://host/api/console
        Authorization: Bearer YOUR_TOKEN
        ```
        
        Send commands as text frames. Responses are streamed back as text frames.
      operationId: console
      security:
        - bearerAuth: []
      responses:
        '101':
          description: WebSocket connection established
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key token. Generate keys via the console command `key create`.
        
        **Permission Levels:**
        - `ALL` — Full admin access to all endpoints
        - `PUBLIC` — Read-only access to public endpoints
...
  responses:
    Unauthorized:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...